+++ app/DataTables/InvoicesTrashDataTable.php
@@ -23,8 +23,9 @@ class InvoicesTrashDataTable extends InvoicesDataTable
             })
             ->editColumn('invoice_status_id', function (Invoice $invoice) use ($statuses) {
                 $ret = '<td class="hidden-sm hidden-xs">
-                <span class="badge badge-' . strtolower($statuses[$invoice->status_text]) . '">
-                    ' . trans('bt.' . strtolower($statuses[$invoice->status_text])) . '</span>';
+                <span class="badge badge-' . strtolower($invoice->status_text) . '">
+                    ' . $statuses[$invoice->status_text] . '</span>';
+
                 if ($invoice->viewed)
                     $ret .= '<span class="badge badge-success">' . trans('bt.viewed') . '</span>';
                 else