+++ database/migrations/2021_11_01_000000_version_533.php
@@ -1,0 +1,30 @@
+<?php
+
+use BT\Modules\Settings\Models\Setting;
+use Illuminate\Database\Migrations\Migration;
+use Illuminate\Database\Schema\Blueprint;
+
+class Version533 extends Migration
+{
+
+    /**
+     * Run the migrations.
+     * @table payments_custom
+     *
+     * @return void
+     */
+    public function up()
+    {
+        Setting::saveByKey('version', '5.3.3');
+    }
+
+    /**
+     * Reverse the migrations.
+     *
+     * @return void
+     */
+     public function down()
+     {
+         Setting::saveByKey('version', '5.3.2');
+     }
+}